# secp256k1: Optimize field inverse calc.