@davecgh 2024-08-12
Rebased to latest master.